Update README.md
Escort Appointment Calendar is a user-friendly WordPress plugin that enables escorts to manage their schedules effortlessly. It allows for easy creation, editing, and updating of appointments with details such as date, time, and location. Clients can book appointments directly through the site, with automatic email notifications for confirmations and changes. Enhance your booking management with this intuitive and responsive plugin.